move tohex to common/src and make it public
